Fix: fix typo in the master_deploy.yml file
RCA:
GitHub Actions configuration file set up in commit: 0c90317e8a763561c8c195d52a61653a33c9f8a2
was copied as it from https://medium.com/@Keithweaver_/setting-up-github-actions-for-a-react-app-on-github-pages-f66b28c312ac.
and removed the configuration portion to test againts multiple node versions but missed that in the name of the step.